Add PNaCl support for VS addin.
The primary changes here are due to the fact that the
link phase for PNaCl differs significantly from from
NaCl one.   Also, since PNaCl toolchains is a bat file
that calls python script, we need to ensure that python
is in the path for all PNaCl builds.

The link phase was using automatic depenedacy tracking
previously but this process fails to launch the python
process successfully, so we now write the .tlog manually
in the same way we do for the compiler phase.

BUG=

Review URL: https://codereview.chromium.org/11266051