[nacltoons] Refactor game data description files.
Pure data files now use the .def extension.

game.def now describes the game and pointers to
the .lua logic for the game.  This is now much
closer to the game design document.

Add validator.lua that checks .def files as they are
read in.

This change is a refactor only and should not effect
the game in any visible way.

Review URL: https://codereview.chromium.org/12742003