[nacltoons] Move scene and menu management to lua
This change does not effect the behaviour of the sample game
but moves more of the game logic into lua code.

It is now up to the game (game.lua) to add any menus, layers,
or flow needed to start, stop and move between levels.
To keep the sample games simple a default game behavior
script contains a standard level selection code.  The idea
is that this can be cloned an modified.

Review URL: https://codereview.chromium.org/12603004